Fikser routing av custompaths i LinkPanelNavno
